+++ src/py42/exceptions.py
@@ -140,6 +140,14 @@ class Py42InvalidRuleOperationError(Py42HTTPError):
         )
 
 
+class Py42MFARequiredError(Py42UnauthorizedError):
+    """An exception raised when a request requires multi-factor authentication"""
+
+    def __init__(self, exception, message=None):
+        message = message or u"User requires multi-factor authentication."
+        super(Py42MFARequiredError, self).__init__(exception, message)
+
+
 def raise_py42_error(raised_error):
     """Raises the appropriate :class:`py42.exceptions.Py42HttpError` based on the given
     HTTPError's response status code.