+++ src/py42/sdk/__init__.py
@@ -32,7 +32,7 @@ from py42.services.users import UserService
 from py42.usercontext import UserContext
 
 
-def from_local_account(host_address, username, password):
+def from_local_account(host_address, username, password, totp=None):
     """Creates a :class:`~py42.sdk.SDKClient` object for accessing the Code42 REST APIs using the
     supplied credentials. Currently, only accounts created within the Code42 console or using the
     APIs (including py42) are supported. Username/passwords that are based on Active Directory,
@@ -43,11 +43,13 @@ def from_local_account(host_address, username, password):
             console.us.code42.com
         username (str): The username of the authenticating account.
         password (str): The password of the authenticating account.
+        totp (str): The time-based one-time password of the authenticating account. Include only
+            if the account uses Code42's two-factor authentication.
 
     Returns:
         :class:`py42.sdk.SDKClient`
     """
-    client = SDKClient.from_local_account(host_address, username, password)
+    client = SDKClient.from_local_account(host_address, username, password, totp)
 
     # test credentials
     client.users.get_current()
@@ -61,7 +63,7 @@ class SDKClient(object):
         self._user_ctx = user_ctx
 
     @classmethod
-    def from_local_account(cls, host_address, username, password):
+    def from_local_account(cls, host_address, username, password, totp=None):
         """Creates a :class:`~py42.sdk.SDKClient` object for accessing the Code42 REST APIs using
         the supplied credentials. Currently, only accounts created within the Code42 console or
         using the APIs (including py42) are supported. Username/passwords that are based on Active
@@ -72,7 +74,8 @@ class SDKClient(object):
                 console.us.code42.com
             username (str): The username of the authenticating account.
             password (str): The password of the authenticating account.
-
+            totp (str): The time-based one-time password of the authenticating account. Include only
+                if the account uses Code42's two-factor authentication.
         Returns:
             :class:`py42.sdk.SDKClient`
         """
@@ -80,6 +83,8 @@ class SDKClient(object):
         if username and password:
             basic_auth = HTTPBasicAuth(username, password)
         auth_connection = Connection.from_host_address(host_address, auth=basic_auth)
+        if totp:
+            auth_connection._headers.update({"totp-auth": str(totp)})
         v3_auth = V3Auth(auth_connection)
         main_connection = Connection.from_host_address(host_address, auth=v3_auth)