+++ CHANGELOG.md
@@ -17,6 +17,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ### Added
 
 - `sdk.alerts.update_state()` method to update state.
+- Support for two-factor authentication in `sdk.from_local_account()`
 
 ## 1.8.1 - 2020-08-28